final minor changes

+++ bindsnet/learning/reward.py
@@ -200,8 +200,8 @@ class DynamicDopamineInjection(AbstractReward):
                 if self.give_reward:
                     if self.label == kwargs['pred_label']:
                         self.dopamine = self.rew_base
-                        self.rew_base =  torch.clip(torch.tensor([self.rew_base - self.dps/self.dps_factor]),  min=self.dps/self.dps_factor,max=self.dps*self.dps_factor)
-                        self.punish_base = torch.clip(torch.tensor([self.punish_base + self.neg_dps/self.dps_factor]),  min=self.neg_dps/self.dps_factor,max=self.neg_dps*self.dps_factor)
+                        self.rew_base =  float(torch.clip(torch.tensor([self.rew_base - self.dps/self.dps_factor]),  min=self.dps/self.dps_factor,max=self.dps*self.dps_factor))
+                        self.punish_base = float(torch.clip(torch.tensor([self.punish_base + self.neg_dps/self.dps_factor]),  min=self.neg_dps/self.dps_factor,max=self.neg_dps*self.dps_factor))
                     else:
                         self.dopamine = -self.punish_base
                         self.rew_base =  float(torch.clip(torch.tensor([self.rew_base + self.dps/self.dps_factor]),  min=self.dps/self.dps_factor,max=self.dps*self.dps_factor))
